refactor(cx.ksim.mather): add `cli` module
I'm planning to add a `gui` version so I just add the foundation
