+++ compile.sh
@@ -1,4 +1,5 @@
 #/usr/bin/env bash
 
 rm -rf bin
-javac -d bin/main --module cx.ksim.mather --module-source-path src/main/
+javac -d bin/main --module cx.ksim.mather.cli --module-source-path src/main/
+javac -d bin/test -cp bin/main/cx.ksim.mather.cli/ src/test/cx.ksim.mather.cli/cx/ksim/mather/test/LexerTest.java src/test/cx.ksim.mather.cli/cx/ksim/mather/test/ParserTest.java
