+++ src/main/cx.ksim.mather.cli/cx/ksim/mather/cli/BinaryExpression.java
@@ -1,4 +1,4 @@
-package cx.ksim.mather;
+package cx.ksim.mather.cli;
 
 public class BinaryExpression implements Expression {
