+++ src/main/cx.ksim.mather.cli/cx/ksim/mather/cli/Expression.java
@@ -1,4 +1,4 @@
-package cx.ksim.mather;
+package cx.ksim.mather.cli;
 
 public interface Expression {
 	abstract double eval();
