+++ src/main/cx.ksim.mather.cli/cx/ksim/mather/cli/FactorToken.java
@@ -1,4 +1,4 @@
-package cx.ksim.mather;
+package cx.ksim.mather.cli;
 
 public enum FactorToken implements TokenKind {
 	NUMBER, OPEN_PAREN, CLOSE_PAREN, UNARY_OP, FUNC_CALL;
