+++ src/main/cx.ksim.mather.cli/cx/ksim/mather/cli/Token.java
@@ -1,3 +1,3 @@
-package cx.ksim.mather;
+package cx.ksim.mather.cli;
 
 public record Token(TokenKind kind, String value) {}
