+++ src/main/cx.ksim.mather.cli/module-info.java
@@ -1,0 +1,3 @@
+module cx.ksim.mather.cli {
+
+}
