--- src/main/cx.ksim.mather/module-info.java
@@ -1,3 +1,0 @@
-module cx.ksim.mather {
-
-}
